refactor: unify event bus and clean chat events
