+++ apps/server/src/handlers/achievementEventHandler.ts
@@ -34,8 +34,6 @@ const ACHIEVEMENT_CHANNELS = [
 
   // Chat events
   REDIS_CHANNELS.CHAT_MESSAGE_SENT,
-  REDIS_CHANNELS.CHAT_TYPING_START,
-  REDIS_CHANNELS.CHAT_TYPING_STOP,
 
   // Leaderboard events
   REDIS_CHANNELS.LEADERBOARD_RANK_UPDATE,
